This commit was generated by cvs2svn to compensate for changes in r5807, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@5808 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/serial.h
@@ -100,7 +100,7 @@ extern serial_t *serial_get_device(int number);
 extern void serial_reset(void);
 
 extern int parallelattention(int b);
-extern int parallelsendbyte(int data);
+extern int parallelsendbyte(BYTE data);
 extern int parallelreceivebyte(BYTE *data, int fake);
 extern void serialattention(void);
 extern void serialsendbyte(void);
